@dcarley 2018-01-02
> You have a spurious ' in the awk arguments within the loop (both in the PR and the commit message).

Amended now. I think it was a copy/paste error.